requirements.txt
streamlit==1.36.0
google-generativeai
python-dotenv==1.0.1
IPython==8.26.0
langchain==0.2.6
PyPDF2==3.0.1
chromadb==0.5.3